requirements.txt
PyQt6==6.6.1
Pillow==10.1.0
piexif==1.1.3
folium==0.14.0
geopy==2.3.0
